fix: split error handling code to separate module
