+++ libs/coin-modules/coin-evm/src/network/node/rpc.common.ts
@@ -14,6 +14,7 @@ import { GasEstimationError, InsufficientFunds, UnsupportedRpcMethodError } from
 import { getSerializedTransaction } from "../../transaction";
 import { FeeHistory } from "../../types";
 import { safeEncodeEIP55, normalizeAddress } from "../../utils";
+import { hasErrorCode, isUnsupportedRpcMethodError } from "./rpc.errors";
 import {
   NodeApi,
   isExternalNodeConfig,
@@ -398,61 +399,6 @@ export const getBlockReceipts: Exclude<NodeApi["getBlockReceipts"], undefined> =
       };
     });
   });
-
-function isUnsupportedRpcMethodError(error: unknown): boolean {
-  const unsupportedCodes = new Set(["-32601", "method_not_found"]);
-  return collectRpcErrorCodes(error).some(code => unsupportedCodes.has(code));
-}
-
-function hasErrorCode(error: unknown, code: string): boolean {
-  if (typeof error !== "object" || error === null || !("code" in error)) return false;
-  return Reflect.get(error, "code") === code;
-}
-
-function normalizeRpcErrorCode(code: unknown): string | null {
-  if (typeof code === "number") {
-    return String(code);
-  }
-  if (typeof code === "string") {
-    return code.toLowerCase();
-  }
-  return null;
-}
-
-function collectRpcErrorCodes(error: unknown): string[] {
-  const result = new Set<string>();
-  const visited = new WeakSet<object>();
-
-  const visit = (value: unknown): void => {
-    if (typeof value !== "object" || value === null || visited.has(value)) {
-      return;
-    }
-    visited.add(value);
-
-    for (const [key, field] of Object.entries(value)) {
-      if (key === "code") {
-        const normalizedCode = normalizeRpcErrorCode(field);
-        if (normalizedCode) {
-          result.add(normalizedCode);
-        }
-      }
-
-      // Some providers wrap RPC error payloads in a stringified response body.
-      if (key === "responseBody" && typeof field === "string") {
-        try {
-          visit(JSON.parse(field));
-        } catch {
-          // ignore malformed response body
-        }
-      } else {
-        visit(field);
-      }
-    }
-  };
-
-  visit(error);
-  return Array.from(result);
-}
 
 function isPrefetchedBlockTransaction(value: unknown): value is PrefetchedBlockTransaction {
   return (
