+++ libs/coin-modules/coin-evm/src/network/node/rpc.native.ts
@@ -1,4 +1,4 @@
-/* istanbul ignore file: pure exports, test unecessary */
+/* istanbul ignore file: pure exports, test unnecessary */
 
 /** ⚠️ keep this order of import. @see https://docs.ethers.io/v5/cookbook/react-native/#cookbook-reactnative ⚠️ */
 import "react-native-get-random-values";
